Move internal usages off of the deprecated classes.
