+++ rx-netty-contexts/src/main/java/io/reactivex/netty/contexts/ContextPipelineConfigurators.java
@@ -23,7 +23,7 @@ import io.reactivex.netty.protocol.http.client.HttpClientRequest;
 import io.reactivex.netty.protocol.http.client.HttpClientResponse;
 import io.reactivex.netty.protocol.http.server.HttpServerRequest;
 import io.reactivex.netty.protocol.http.server.HttpServerResponse;
-import io.reactivex.netty.protocol.text.sse.ServerSentEvent;
+import io.reactivex.netty.protocol.http.sse.ServerSentEvent;
 
 /**
  * A factory class for different {@link PipelineConfigurator} for the context module.
