+++ rx-netty-examples/src/main/java/io/reactivex/netty/examples/http/sse/HttpSseServer.java
@@ -23,7 +23,7 @@ import io.reactivex.netty.protocol.http.server.HttpServer;
 import io.reactivex.netty.protocol.http.server.HttpServerRequest;
 import io.reactivex.netty.protocol.http.server.HttpServerResponse;
 import io.reactivex.netty.protocol.http.server.RequestHandler;
-import io.reactivex.netty.protocol.text.sse.ServerSentEvent;
+import io.reactivex.netty.protocol.http.sse.ServerSentEvent;
 import rx.Notification;
 import rx.Observable;
 import rx.functions.Func1;
