+++ rx-netty/src/test/java/io/reactivex/netty/protocol/http/client/HttpClientTest.java
@@ -34,7 +34,7 @@ import io.reactivex.netty.pipeline.PipelineConfigurator;
 import io.reactivex.netty.pipeline.PipelineConfigurators;
 import io.reactivex.netty.protocol.http.server.HttpServer;
 import io.reactivex.netty.protocol.http.server.HttpServerBuilder;
-import io.reactivex.netty.protocol.text.sse.ServerSentEvent;
+import io.reactivex.netty.protocol.http.sse.ServerSentEvent;
 import io.reactivex.netty.server.RxServerThreadFactory;
 import org.junit.AfterClass;
 import org.junit.Assert;
